add login and logout, add search

+++ base/views.py
@@ -1,22 +1,91 @@
 from django.shortcuts import render, redirect
-from .models import Room
+from django.http import HttpResponse
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
+from .models import Room, Topic, Message
 from .forms import RoomForm
 
 # Create your views here.
 
 
+def loginPage(request):
+    page = "login"
+    if request.user.is_authenticated:
+        return redirect("home")
+
+    if request.method == "POST":
+        username = request.POST.get("username").lower()
+        password = request.POST.get("password")
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, "User does not exist")
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect("home")
+        else:
+            messages.error(request, "Username OR password does not exist")
+    context = {"page": page}
+    return render(request, "base/login_register.html", context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect("home")
+
+
+def registerPage(request):
+    form = UserCreationForm()
+    if request.method == "POST":
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect("home")
+        else:
+            messages.error(request, "An error occured during registration")
+    context = {"form": form}
+    return render(request, "base/login_register.html", context)
+
+
 def home(request):
-    rooms = Room.objects.all()
-    context = {"rooms": rooms}
+    q = request.GET.get("q") if request.GET.get("q") != None else ""
+
+    rooms = Room.objects.filter(
+        Q(topic__name__icontains=q) | Q(name__icontains=q) | Q(description__icontains=q)
+    )
+    room_count = rooms.count()
+    topics = Topic.objects.all()
+    context = {"rooms": rooms, "topics": topics, "room_count": room_count}
     return render(request, "base/home.html", context)
 
 
 def room(request, pk):
     room = Room.objects.get(id=pk)
-    context = {"room": room}
+    # Child model olan Message'ın kayıtlarını getirir.
+    room_messages = room.message_set.all().order_by("-created")
+
+    if request.method == "POST":
+        message = Message.objects.create(
+            user=request.user, room=room, body=request.POST.get("body")
+        )
+        return redirect("room", pk=room.id)
+
+    context = {"room": room, "room_messages": room_messages}
     return render(request, "base/room.html", context)
 
 
+@login_required(login_url="login")
 def createRoom(request):
     form = RoomForm()
 
@@ -30,9 +99,13 @@ def createRoom(request):
     return render(request, "base/room_form.html", context)
 
 
+@login_required(login_url="login")
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!!")
 
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
@@ -44,8 +117,12 @@ def updateRoom(request, pk):
     return render(request, "base/room_form.html", context)
 
 
+@login_required(login_url="login")
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!!")
 
     if request.method == "POST":
         room.delete()
